feat(decouple): lift the routing brain into the engine (Move 4)
Move the probabilistic body of apps/credence-pi/brain/routing_brain.jl into
src/routing.jl (inside module Ontology, after structure_bma.jl): RoutingState,
EmissionBelief, LatencyBelief, the named decision templates route/route_eu/
escalation_next/posterior_accuracy, the coupled-EM route_outcome!, the
closed-form decode_correctness, route_decide/escalate_decide (un-underscored
from the app's _route_decide/_escalate_decide), and the data-in
reconstruct_*_from_data. Composes only Tier-1 ops (optimise, condition on
Beta×WeightedBernoulli, structure_observe_soft, ProductMeasure,
LinearCombination/Projection) — no new frozen type, no new axiom-constrained fn.
Routing reuses the already-lifted StructureBMA substrate.

routing_brain.jl collapses to a thin shim (like feature_brain.jl): re-exports
the lifted names, keeps the env-reading wire_routing! + the path-reading
reconstruction wrappers (the embedding daemon reads its committed counts; the
wire path will ship them as inline data). No optimise/LinearCombination/
Projection/ProductMeasure import remains — the leak-closed witness.

decode_correctness stays closed-form (θ̄ at the mixture level via
posterior_accuracy, not a per-cell α/(α+β)) — the only way the emission M-step
stays weighted by the coherent per-turn correctness and replay stays bit-exact.

Verified: test_routing.jl passes byte-for-byte unmodified (40 assertions incl.
the exact-replay oracle); test_server.jl (23), test_structure_bma, test_core
green; every eval/ + daemon RoutingBrain reach-in resolves; lint clean (226).
