+++ src/ontology.jl
@@ -1752,4 +1752,10 @@ export StructureBMA, build_structure_model, build_structure_prior,
        belief_at_context, context_from_features, structure_firing_tags,
        structure_decision_kernel
 
+include("routing.jl")
+export RoutingState, EmissionBelief, LatencyBelief, route, route_eu, escalation_next,
+       posterior_accuracy, route_outcome!, decode_correctness, latency_at,
+       route_decide, escalate_decide, reconstruct_latency_from_data,
+       reconstruct_routing_tops_from_data, _ctx_key
+
 end # module Ontology
